Test Console
I created this HTML/JS test console to print to the screen instead of
having to pull up the console every time. This helps in my native
environment because pulling up the Chrome Console disconnects the
environment and requires you to reconnect it each time you want to see
an update. Now I just have to refresh.